Fix for appliance build.

+++ src/arpmgrd.c
@@ -1588,8 +1588,8 @@ all_neighbors_len %d\n",nbr_alloc_cnt,
                 port_data_alloc,ovsrec_neighbor_alloc_cnt,
                 idl_nbr_force_ins_cnt,idl_nbr_force_del_cnt,ovs_idl_nbr_cnt,
                 all_neighbors_len);
-    ds_put_format(&ds,"size_neighbor_data = %u, size_ovsrec_neighbor = %u\n",sizeof(struct neighbor_data),
-    sizeof(struct ovsrec_neighbor));
+    ds_put_format(&ds,"size_neighbor_data = %u, size_ovsrec_neighbor = %u\n",(unsigned int)sizeof(struct neighbor_data),
+    (unsigned int)sizeof(struct ovsrec_neighbor));
      ds_put_format(&ds,"nl_dump_req_cnt %d, nl_dump_res_cnt %d, not_multi_part_dump_res_cnt %d \n\
 nl_probe_req_cnt %d, nl_dump_res_size %d,ovsdb_tr_trigger_cnt %d \n",
 nl_dump_req_cnt,nl_dump_res_cnt, nl_dump_res_nof_multi_cnt,
